requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66
matplotlib>=3.7
seaborn>=0.13
kaggle>=1.6
PyYAML>=6.0